+++ frontend/index.jsx
@@ -1,27 +1,10 @@
 import * as React from 'react';
 import ReactDOM from 'react-dom';
-import { App } from './App';
+import Themer from './Themer';
 import { HashRouter as Router } from 'react-router-dom';
-import { createMuiTheme } from '@material-ui/core';
-import { green } from '@material-ui/core/colors';
-import { ThemeProvider } from '@material-ui/core/styles';
-
-const theme = createMuiTheme({
-  palette: {
-    primary: {
-      main: '#FA8756',
-      contrastText: '#ffffff',
-    },
-    secondary: green,
-    type: 'light',
-  },
-});
-
 ReactDOM.render(
   <Router>
-    <ThemeProvider theme={theme}>
-      <App />
-    </ThemeProvider>
+    <Themer />
   </Router>,
   document.getElementById('root')
 );
